Setup new frontend structure with multi page folders
