+++ src/app/components/Background.js
@@ -1,0 +1,1 @@
+// The background image shared across all pages, not sure if this should be a component or be embedded in the styles
