+++ src/app/components/Button.js
@@ -1,0 +1,2 @@
+// This will be our custom generic button Component. Should display a text with an optional icon.
+// Reference: "Generate Cover letter" Button
