+++ src/app/components/Divider.js
@@ -1,0 +1,1 @@
+// This is the generic divider "------------- OR -------------" that I'm going to use everywhere!
