+++ src/app/components/NavBar.js
@@ -1,0 +1,2 @@
+// This will be the upper right navigation bar used in every page. Should have responsive flexing that adjust to the size of
+// the navbar contents "Home, Dashboard, Logout, etc". The buttons should also be clickable and navigate to the respective routes
