+++ src/app/components/TextArea.js
@@ -1,0 +1,2 @@
+// This will be our custom generic TextArea component. Should be styled with the green design and accept optinal placeholder text.
+// Reference: "Paste the job description here.." TextArea
