+++ src/app/components/Title.js
@@ -1,0 +1,2 @@
+// This will be the center large "SpeedFill" title used in most page. Should have an option to display the subtitles or not.
+// Some pages may only want to display the title without the subtitles.
