+++ src/app/hooks/useAIPrompt.js
