+++ src/app/pages/form/components/Container.js
@@ -1,0 +1,5 @@
+// This is the floating green container in the form page. Should include a title and subtitle, the children can be passed
+// into the component rather than hardcoding them in the component itself.
+
+// For example, the "Your Resume" component will create a container with said title and subtitle, then the
+// higher level component will pass in the dropzone as well as the text area in the children prop
