+++ src/app/pages/form/components/Dropzone.js
@@ -1,0 +1,1 @@
+// The dropzone component in the resume input form
