+++ src/app/styles/theme.js
@@ -1,0 +1,19 @@
+"use client";
+
+import { Roboto } from "next/font/google";
+import { createTheme } from "@mui/material/styles";
+
+const roboto = Roboto({
+  weight: ["300", "400", "500", "700"],
+  subsets: ["latin"],
+  display: "swap",
+});
+
+const theme = createTheme({
+  cssVariables: true,
+  typography: {
+    fontFamily: roboto.style.fontFamily,
+  },
+});
+
+export default theme;
